+++ test/noahtheduke/splint/netrunner_test.clj
@@ -27,6 +27,7 @@
     lint/let-when 1
     lint/redundant-call 3
     lint/thread-macro-one-arg 71
+    lint/warn-on-reflection 220
     metrics/fn-length 485
     metrics/parameter-count 204
     naming/conventional-aliases 33
@@ -77,4 +78,4 @@
              :diagnostics
              (group-by :rule-name)
              (#(update-vals % count)))))
-    (expect 1451 (count (:diagnostics results)))))
+    (expect 1671 (count (:diagnostics results)))))