+++ test/noahtheduke/splint/re_frame_test.clj
@@ -18,6 +18,7 @@
   '{lint/if-else-nil 2
     lint/thread-macro-one-arg 15
     lint/try-splicing 1
+    lint/warn-on-reflection 32
     metrics/fn-length 20
     naming/conventional-aliases 2
     naming/predicate 1
@@ -42,4 +43,4 @@
              :diagnostics
              (group-by :rule-name)
              (#(update-vals % count)))))
-    (expect 51 (count (:diagnostics results)))))
+    (expect 83 (count (:diagnostics results)))))